+++ tests/__init__.py
@@ -1,1 +1,7 @@
 """ Tests """
+import six
+
+if six.PY3:  # pragma: no cover
+    import unittest
+    # pylint: disable=E1101
+    unittest.TestCase.assertItemsEqual = unittest.TestCase.assertCountEqual