+++ tests/test_params.py
@@ -1,18 +1,18 @@
 """ Tests for param utilities """
 from __future__ import unicode_literals
 
-import calendar
 import datetime
-import json
 import time
 
+import calendar
+import json
+import six
 from mock import MagicMock, call, patch
 from pyramid.config import Configurator
 from pyramid.httpexceptions import HTTPBadRequest
 from pyramid.testing import DummyRequest
 
 import pyramid_duh
-from pyramid_duh.compat import is_bytes, is_string, string_type
 from pyramid_duh.params import argify, param, includeme
 
 
@@ -73,7 +73,7 @@ class TestParam(unittest.TestCase):
         request.params = {'field': 'myfield'}
         field = param(request, 'field')
         self.assertEquals(field, 'myfield')
-        self.assertTrue(is_string(field, strict=True))
+        self.assertTrue(isinstance(field, six.text_type))
 
     def test_unicode_json_body(self):
         """ Pull unicode params out of json body """
@@ -83,22 +83,22 @@ class TestParam(unittest.TestCase):
         request.headers = {'Content-Type': 'application/json'}
         field = param(request, 'field')
         self.assertEquals(field, 'myfield')
-        self.assertTrue(is_string(field, strict=True))
+        self.assertTrue(isinstance(field, six.text_type))
 
     def test_unicode_param_explicit(self):
         """ Specifying type=unicode checks arg type before returning it """
         request = DummyRequest()
         request.params = {'field': 'myfield'}
-        field = param(request, 'field', type=string_type)
+        field = param(request, 'field', type=six.text_type)
         self.assertEquals(field, 'myfield')
-        self.assertTrue(is_string(field, strict=True))
+        self.assertTrue(isinstance(field, six.text_type))
 
     def test_unicode_param_bad_type(self):
         """ Raise exception if unicode param as incorrect type """
         request = DummyRequest()
         request.params = {'field': 4}
         with self.assertRaises(HTTPBadRequest):
-            param(request, 'field', type=string_type)
+            param(request, 'field', type=six.text_type)
 
     def test_str_param(self):
         """ Pull binary string param off of request object """
@@ -106,7 +106,7 @@ class TestParam(unittest.TestCase):
         request.params = {'field': 'myfield'}
         field = param(request, 'field', type=bytes)
         self.assertEquals(field, b'myfield')
-        self.assertTrue(is_bytes(field))
+        self.assertTrue(isinstance(field, six.binary_type))
 
     def test_str_json_body(self):
         """ Pull str params out of json body """
@@ -116,7 +116,7 @@ class TestParam(unittest.TestCase):
         request.headers = {'Content-Type': 'application/json'}
         field = param(request, 'field', type=bytes)
         self.assertEquals(field, b'myfield')
-        self.assertTrue(is_bytes(field))
+        self.assertTrue(isinstance(field, six.binary_type))
 
     def test_bytes_param(self):
         """ Pull binary string param off of request object """
@@ -124,7 +124,7 @@ class TestParam(unittest.TestCase):
         request.params = {'field': 'myfield'}
         field = param(request, 'field', type=bytes)
         self.assertEquals(field, b'myfield')
-        self.assertTrue(is_bytes(field))
+        self.assertTrue(isinstance(field, six.binary_type))
 
     def test_int_param(self):
         """ Pull integer off of request object """
@@ -480,7 +480,7 @@ class TestArgify(unittest.TestCase):
         """ argify() can run a validation check on parameter values """
         validate = lambda x: x.startswith('foo')
 
-        @argify(field=(string_type, validate))
+        @argify(field=(six.text_type, validate))
         def base_req(request, field):
             return field
         context = object()
@@ -493,7 +493,7 @@ class TestArgify(unittest.TestCase):
         """ if argify fails validation check, raise exception """
         validate = lambda x: x.startswith('foo')
 
-        @argify(field=(string_type, validate))
+        @argify(field=(six.text_type, validate))
         def base_req(request, field):  # pragma: no cover
             return field
         context = object()
@@ -506,7 +506,7 @@ class TestArgify(unittest.TestCase):
         """ if argify fails validation check for kwargs, raise exception """
         validate = lambda x: x.startswith('foo')
 
-        @argify(field=(string_type, validate))
+        @argify(field=(six.text_type, validate))
         def base_req(request, field=None):  # pragma: no cover
             return field
         context = object()