+++ doc/conf.py
@@ -13,8 +13,6 @@ addtl_paths = (
 for path in addtl_paths:
     sys.path.insert(0, os.path.abspath(os.path.join(docs_basepath, path)))
 
-from pyramid_duh_version import git_version_data
-
 extensions = ['sphinx.ext.autodoc', 'numpydoc', 'sphinx.ext.intersphinx',
               'sphinx.ext.linkcode', 'sphinx.ext.autosummary']
 
@@ -22,9 +20,8 @@ master_doc = 'index'
 project = u'pyramid_duh'
 copyright = u'2013, Steven Arcangeli'
 
-version_data = git_version_data()
-version = version_data['tag']
-release = version_data['version']
+release = '0.1.2'
+version = '.'.join(release.split('.')[:2])
 
 exclude_patterns = ['_build']
 pygments_style = 'sphinx'
@@ -42,4 +39,4 @@ def linkcode_resolve(domain, info):
     if domain != 'py' or not info['module']:
         return None
     filename = info['module'].replace('.', '/')
-    return "https://github.com/stevearc/pyramid_duh/blob/%s/%s.py" % (version_data['ref'], filename)
+    return "https://github.com/stevearc/pyramid_duh/blob/%s/%s.py" % (release, filename)