+++ CHANGES.rst
@@ -3,6 +3,7 @@
 * Bug fix: IStaticResource fails to look up self.request if nested 2-deep
 * Bug fix: Name collisions with version_helper.py
 * Bug fix: Subpath glob matching always respects case
+* Feature: @argify works on view classes
 
 0.1.0
 -----