+++ CHANGES.rst
@@ -1,3 +1,11 @@
+Changelog
+=========
+
+0.1.2
+-----
+* Bug fix: Fix potential timezone issue when converting unix time to datetime
+* Using the 'six' library for python 2/3 compatibility
+
 0.1.1
 -----
 * Bug fix: IStaticResource fails to look up self.request if nested 2-deep